+++ the_warroom/models.py
@@ -1640,7 +1640,7 @@ class Game(models.Model):
     hirelings = models.ManyToManyField(Hireling, blank=True, related_name='games')
     undrafted_faction = models.ForeignKey(Faction, on_delete=models.PROTECT, null=True, blank=True, default=None, related_name='undrafted_games')
     undrafted_vagabond = models.ForeignKey(Vagabond, on_delete=models.PROTECT, null=True, blank=True, default=None, related_name='undrafted_games')
-    undrafted_captains = models.ForeignKey(Vagabond, on_delete=models.PROTECT, null=True, blank=True, default=None, related_name='undrafted_captain_games')
+    undrafted_captains = models.ManyToManyField(Vagabond, blank=True, related_name='undrafted_captain_games')
     link = models.URLField(max_length=1000, null=True, blank=True)
     video_link = models.URLField(
         max_length=500,
@@ -1838,6 +1838,7 @@ class Effort(models.Model):
     faction = models.ForeignKey(Faction, on_delete=models.PROTECT, related_name='efforts', blank=True, null=True)
     vagabond = models.ForeignKey(Vagabond, on_delete=models.PROTECT, null=True, blank=True, default=None, related_name='efforts')
     captains = models.ManyToManyField(Vagabond, blank=True, related_name='efforts_as_captain')
+    discarded_captain = models.ForeignKey(Vagabond, on_delete=models.PROTECT, null=True, blank=True, default=None, related_name='discarded_efforts')
     coalition_with = models.ForeignKey(Faction, on_delete=models.PROTECT, null=True, blank=True, related_name='efforts_in_coalition')
     dominance = models.CharField(max_length=20, choices=DominanceChoices.choices, null=True, blank=True)
     brazen_demogogue = models.BooleanField(default=False)